+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # unreleased
 
+* AET artifacts deployed to Karaf *deploy* folder
 * Fix for Karaf cache clearing
 
 # 3.0.0